refactor: rename and move types and utils
